requirements.txt
GitPython>=3.0.2
prompter>=0.3.10
python-dateutil>=2.8.0
pytest>=5.1.1
pyyaml>=5.1.2